SCRUM-1511 refactored code
