+++ src/main/cliapp/src/containers/reportsPage/ReportDialog/ReportDialog.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Dialog } from 'primereact/dialog';
 import { Button } from 'primereact/button';
 import { DialogTop } from './DialogTop';
